requirement.txt
# how to install in bulk
# pip install -r requirement.txt

streamlit
mysql.connector